refactor cli commands
- incorporate `cluster` command
- re-org `bitcoin` and `ln` into subcommands
