+++ src/warnet/cli/scenarios.py
@@ -12,7 +12,7 @@ from .rpc import rpc_call
 
 @click.group(name="scenarios")
 def scenarios():
-    """Scenario commands"""
+    """Manage scenarios on a running network"""
 
 
 @scenarios.command()
