+++ test/ln_test.py
@@ -40,7 +40,7 @@ class LNTest(TestBase):
 
     def run_ln_init_scenario(self):
         self.log.info("Running LN Init scenario")
-        self.warcli("rpc 0 getblockcount")
+        self.warcli("bitcoin rpc 0 getblockcount")
         self.warcli("scenarios run ln_init")
         self.wait_for_all_scenarios()
         scenario_return_code = self.get_scenario_return_code("ln_init")
@@ -49,9 +49,9 @@ class LNTest(TestBase):
 
     def test_channel_policies(self):
         self.log.info("Ensuring node-level channel policy settings")
-        node2pub, node2host = json.loads(self.warcli("lncli 2 getinfo"))["uris"][0].split("@")
-        chan_id = json.loads(self.warcli("lncli 2 listchannels"))["channels"][0]["chan_id"]
-        chan = json.loads(self.warcli(f"lncli 2 getchaninfo {chan_id}"))
+        node2pub, node2host = json.loads(self.warcli("ln rpc 2 getinfo"))["uris"][0].split("@")
+        chan_id = json.loads(self.warcli("ln rpc 2 listchannels"))["channels"][0]["chan_id"]
+        chan = json.loads(self.warcli(f"ln rpc 2 getchaninfo {chan_id}"))
 
         # node_1 or node_2 is tank 2 with its non-default --bitcoin.timelockdelta=33
         if chan["node1_policy"]["time_lock_delta"] != 33:
@@ -64,15 +64,15 @@ class LNTest(TestBase):
 
     def test_ln_payment_0_to_2(self):
         self.log.info("Test LN payment from 0 -> 2")
-        inv = json.loads(self.warcli("lncli 2 addinvoice --amt=2000"))["payment_request"]
+        inv = json.loads(self.warcli("ln rpc 2 addinvoice --amt=2000"))["payment_request"]
         self.log.info(f"Got invoice from node 2: {inv}")
         self.log.info("Paying invoice from node 0...")
-        self.log.info(self.warcli(f"lncli 0 payinvoice -f {inv}"))
+        self.log.info(self.warcli(f"ln rpc 0 payinvoice -f {inv}"))
 
         self.wait_for_predicate(self.check_invoice_settled)
 
         self.log.info("Ensuring channel-level channel policy settings: source")
-        payment = json.loads(self.warcli("lncli 0 listpayments"))["payments"][0]
+        payment = json.loads(self.warcli("ln rpc 0 listpayments"))["payments"][0]
         assert (
             payment["fee_msat"] == "5506"
         ), f"Expected fee_msat to be 5506, got {payment['fee_msat']}"
@@ -82,22 +82,22 @@ class LNTest(TestBase):
 
     def test_ln_payment_2_to_0(self):
         self.log.info("Test LN payment from 2 -> 0")
-        inv = json.loads(self.warcli("lncli 0 addinvoice --amt=1000"))["payment_request"]
+        inv = json.loads(self.warcli("ln rpc 0 addinvoice --amt=1000"))["payment_request"]
         self.log.info(f"Got invoice from node 0: {inv}")
         self.log.info("Paying invoice from node 2...")
-        self.log.info(self.warcli(f"lncli 2 payinvoice -f {inv}"))
+        self.log.info(self.warcli(f"ln rpc 2 payinvoice -f {inv}"))
 
         self.wait_for_predicate(lambda: self.check_invoices(0) == 1)
 
         self.log.info("Ensuring channel-level channel policy settings: target")
-        payment = json.loads(self.warcli("lncli 2 listpayments"))["payments"][0]
+        payment = json.loads(self.warcli("ln rpc 2 listpayments"))["payments"][0]
         assert (
             payment["fee_msat"] == "2213"
         ), f"Expected fee_msat to be 2213, got {payment['fee_msat']}"
 
     def test_simln(self):
         self.log.info("Engaging simln")
-        node2pub, _ = json.loads(self.warcli("lncli 2 getinfo"))["uris"][0].split("@")
+        node2pub, _ = json.loads(self.warcli("ln rpc 2 getinfo"))["uris"][0].split("@")
         activity = [
             {"source": "ln-0", "destination": node2pub, "interval_secs": 1, "amount_msat": 2000}
         ]
@@ -109,14 +109,14 @@ class LNTest(TestBase):
         assert self.check_invoices(1) == 0, "Expected no invoices for node 1"
 
     def check_invoice_settled(self):
-        invs = json.loads(self.warcli("lncli 2 listinvoices"))["invoices"]
+        invs = json.loads(self.warcli("ln rpc 2 listinvoices"))["invoices"]
         if len(invs) > 0 and invs[0]["state"] == "SETTLED":
             self.log.info("Invoice settled")
             return True
         return False
 
     def check_invoices(self, index):
-        invs = json.loads(self.warcli(f"lncli {index} listinvoices"))["invoices"]
+        invs = json.loads(self.warcli(f"ln rpc {index} listinvoices"))["invoices"]
         settled = sum(1 for inv in invs if inv["state"] == "SETTLED")
         self.log.debug(f"Node {index} has {settled} settled invoices")
         return settled
